changement getImage

See merge request leroyj/tower-defense!9

+++ src/model/Coordinates.java
@@ -1,18 +1,18 @@
 package model;
 public class Coordinates {
 
-    private int x,y;
+    private float x,y;
 
-    public Coordinates(int x, int y){
+    public Coordinates(float x, float y){
         this.x=x;
         this.y=y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 }
